[WIP]Add AlphaBeta pruning

+++ src/util.rs
@@ -1,7 +1,8 @@
 // These may be faster than "for" statement.
 // On my computer, these were about 8 times faster than "for" statement in 0.5 billion loops.
 
-pub fn loop_n<F: FnMut()>(n: i32, mut f: F) {
+#[inline(always)]
+pub fn loop_n<F: FnMut()>(n: u64, mut f: F) {
     let mut i = 0;
     loop {
         f();
@@ -12,7 +13,8 @@ pub fn loop_n<F: FnMut()>(n: i32, mut f: F) {
     }
 }
 
-pub fn for_n<F: FnMut(i32)>(n: i32, mut f: F) {
+#[inline(always)]
+pub fn for_n<F: FnMut(u64)>(n: u64, mut f: F) {
     let mut i = 0;
     loop {
         f(i);
